+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Added `list-local-files` command.
+- Support adding custom emojis to profile description.
 
 ### Fixed